third_party/blink/web_tests/http/tests/inspector-protocol/animation/animation-scroll-timeline-expected.txt

Tests that the viewOrScrollTimeline is included for scroll driven animations
PASS: startTime is reported correctly as percentage.
PASS: source.duration is reported correctly as percentage.
PASS: Axis of view or scroll timeline is correct.
PASS: startOffset exists.
PASS: endOffset exists.
PASS: sourceNodeId exists.
